working on class logging mixin
